requirements.txt
Django~=5.0.3
djangorestframework~=3.15.1
alembic~=1.13.1
SQLAlchemy~=2.0.29
django-filter~=24.2
mysqlclient~=2.2.4